Copy message headers (inc correlationId) from first message
In the case of request-reponse, this is what is required (there is
only one request and one response). If the request results in a
stream being produced the gateway won't care that about the
2nd and susequent messages, and it will worry about any it sees
with a correlationId that it can't match with an input.

This change takes the message headers from the each request
and copies them to the first subsequent response.