Add some tests for loading functions from jars (fat and not fat)
Uses the spring-boot-loader to extract a classpath from a fat jar
if it looks like it has BOOT-INF packaging. Otherwise assumes the
jar is flat (e.g. shaded). The test is a good start, but we should
add more because there is a missing feature - @Beans (rather than
@Components or scanned functions) will not have enough type
information in their runtime class metadata. Probably need to
go into the FunctionInspector in the user context to extract the
type information.

See #31