Make errors in grpc input visible
It is better to throw the exception and let reactor log it than
to try and relay it to the caller. As a side effect, since to test
it we needed an empty input stream, we now have support for
event sources (just write a Supplier and it will be adapted to
a Function that discards its input).