Add a flag so that app can close when stream ends
If user sets grpc.exitOnComplete=true then the app will close
down after the end of the output flux (which is the same as the
end of the input flux for a Consumer).

See https://github.com/projectriff/riff/pull/544 for a change
in the sidecar that supports this, and makes it possible to
declare a k8s Job with a sidecar and a function invoker.