No need to start Tomcat if the protocol is grpc
The java-invoker.yaml has a hard-coded protocol of "grpc" so it makes
no sense to waste resources listening on port 8080. This implementation
relies on an environment variable (or equivalently system property
or command line argument) being passed in by the caller. In riff this
would be set up by the function-controller.